+++ src/Molecules/Modal/Modal.stories.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import { Modal, Typography, Box, Button, Flexbox, Icon } from '../..';
+import { Modal, Typography, Box, Button, Flexbox, Icon, FadedScroll } from '../..';
 
 const ScrollMaker = styled.div`
   background-image: linear-gradient(${p => p.theme.color.positive}, ${p => p.theme.color.negative});
@@ -104,30 +104,60 @@ export const footerStory = () => {
         <ScrollMaker />
         {/* ScrollMaker is just used to show how the Modal locks scrolling when open */}
         <Modal onClose={onClose} title="Dialog information" open={open} footer={footer}>
-          <Box mb={2}>
+          <FadedScroll enableMobileFade>
+            <Box mb={2}>
+              <Typography type="primary" as="p">
+                Modals should be used with care as they are quite intrusive on the user experience
+                and demand immediate attention (while also blocking all other actions on the site).
+                Always consider if you can solve a problem in another way first before you choose to
+                go with the modal. Modals should be used with care as they are quite intrusive on
+                the user experience and demand immediate attention (while also blocking all other
+                actions on the site). Always consider if you can solve a problem in another way
+                first before you choose to go with the modal.
+              </Typography>
+            </Box>
+            <Box mb={2}>
+              <Typography type="primary" as="p">
+                That being said they are a good tool if you need to grab the users attention, either
+                to communicate something very important or make them take an action before
+                proceeding. That being said they are a good tool if you need to grab the users
+                attention, either to communicate something very important or make them take an
+                action before proceeding.
+              </Typography>
+            </Box>
+            <Box mb={2}>
+              <Typography type="primary" as="p">
+                Modals should be used with care as they are quite intrusive on the user experience
+                and demand immediate attention (while also blocking all other actions on the site).
+                Always consider if you can solve a problem in another way first before you choose to
+                go with the modal. Modals should be used with care as they are quite intrusive on
+                the user experience and demand immediate attention (while also blocking all other
+                actions on the site). Always consider if you can solve a problem in another way
+                first before you choose to go with the modal.
+              </Typography>
+            </Box>
+            <Box mb={2}>
+              <Typography type="primary" as="p">
+                Modals should be used with care as they are quite intrusive on the user experience
+                and demand immediate attention (while also blocking all other actions on the site).
+                Always consider if you can solve a problem in another way first before you choose to
+                go with the modal. Modals should be used with care as they are quite intrusive on
+                the user experience and demand immediate attention (while also blocking all other
+                actions on the site). Always consider if you can solve a problem in another way
+                first before you choose to go with the modal.
+              </Typography>
+            </Box>
+            <Box mb={2}>
+              <Typography type="primary" as="p">
+                Nielsen/Norman has an excellent article about their usage here
+              </Typography>
+            </Box>
             <Typography type="primary" as="p">
-              Modals should be used with care as they are quite intrusive on the user experience and
-              demand immediate attention (while also blocking all other actions on the site). Always
-              consider if you can solve a problem in another way first before you choose to go with
-              the modal.
+              <a href="https:// www.nngroup.com/articles/modal-nonmodal-dialog/">
+                https:// www.nngroup.com/articles/modal-nonmodal-dialog/
+              </a>
             </Typography>
-          </Box>
-          <Box mb={2}>
-            <Typography type="primary" as="p">
-              That being said they are a good tool if you need to grab the users attention, either
-              to communicate something very important or make them take an action before proceeding.
-            </Typography>
-          </Box>
-          <Box mb={2}>
-            <Typography type="primary" as="p">
-              Nielsen/Norman has an excellent article about their usage here
-            </Typography>
-          </Box>
-          <Typography type="primary" as="p">
-            <a href="https:// www.nngroup.com/articles/modal-nonmodal-dialog/">
-              https:// www.nngroup.com/articles/modal-nonmodal-dialog/
-            </a>
-          </Typography>
+          </FadedScroll>
         </Modal>
       </div>
     );