+++ src/storyshots.test.ts
@@ -3,8 +3,9 @@ import ReactDOM from 'react-dom';
 
 // eslint-disable-next-line import/no-extraneous-dependencies
 import initStoryshots, { multiSnapshotWithOptions } from '@storybook/addon-storyshots';
+import 'jest-styled-components';
 // eslint-disable-next-line import/no-extraneous-dependencies
-import styleSheetSerializer from 'jest-styled-components/src/styleSheetSerializer';
+import { styleSheetSerializer } from 'jest-styled-components/serializer';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { addSerializer } from 'jest-specific-snapshot';
 
@@ -24,6 +25,10 @@ global.IntersectionObserver = class IntersectionObserver {
   }
 
   unobserve() {
+    return null;
+  }
+
+  disconnect() {
     return null;
   }
 };